requirements.txt
absl-py==1.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
ci-info==0.3.0
click==8.1.3
colorama==0.4.5
configobj==5.0.6
configparser==5.2.0
etelemetry==0.3.0
filelock==3.7.1
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.20.4
idna==3.3
isodate==0.6.1
looseversion==1.0.1
lxml==4.9.1
networkx==2.8.5
nibabel==4.0.1
nipype==1.8.3
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pathlib==1.0.1
Pillow==9.2.0
prov==2.0.0
pydot==1.4.2
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
pyxnat==1.4
rdflib==6.2.0
regex==2022.7.25
requests==2.28.1
scipy==1.9.0
simplejson==3.17.6
six==1.16.0
traits==6.3.2
urllib3==1.26.11
